### 2018-02-01 #564 @piermorel
Thank you @cMadan  it took a while, but now gramm feels complete enough!

As for reviewers, I don't have strong suggestions. I was thinking about [Brian Lau](https://github.com/brian-lau), who is a neuroscientist and develops with Matlab (Matlab interface for Stan) and R. From the provided list, [Tania Allard](https://github.com/trallard) appears to have an interest in both Matlab and dataviz.
