### 2018-02-01 #564 @nickwan
I don't think I can ethically be a reviewer -- pretty sure Pierre is using
one of my papers as an example in it

On Thu, Feb 1, 2018 at 1:13 PM, Christopher Madan <notifications@github.com>
wrote:

> @brian-lau <https://github.com/brian-lau> and @trallard
> <https://github.com/trallard>, would either of you be interested in
> reviewing this submission?
>
> @nickwan <https://github.com/nickwan>, I thought you may also be
> interested, and I think you've used gramm before too?
>
