### 2018-02-01 #564 @cMadan
@nickwan, I think it's fine for you to be a reviewer as long as you didn't help develop the package. The review process here is more about the package being functional, sufficiently documented, and including some specific aspects (e.g., license, contributor guidelines, etc.)
