### 2018-02-01 #564 @trallard
This does sound quite interesting. I would for sure be interested in reviewing this or co-reviewing if anyone else is interested
