### 2018-02-05 #564 @cMadan
@trallard and @brian-lau, I will go ahead and start the review now. I am going to assign @trallard as the reviewer, as the JOSS system currently lets us assign one person as the reviewer, but please both do test the code and provide feedback equally as reviewers. Thank you again for volunteering to review this submission!
