### 2018-02-08 piermorel/gramm#56 @trallard
Installation instructions
Even though the (quite simple) installation instructions work as they are, I feel these a bit a dry especially for people only getting familiar with Git(Hub) and or MATLAB, even though if these seems very straightforward or trivial.

Feel free to disagree with this
