### 2018-02-08 piermorel/gramm#57 @trallard
Testing suite implementation
gramm seems well developed so far but I could not find any test for the package.

I might have missed something, but if I haven't, have you considered adding any?
