### 2018-02-08 #568 @trallard
Currently doing the review of gramm and have a couple of things to raise:

**Documentation**

> - [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Even though the instructions allow for correct installation they might be a bit dry? cryptic? to those that have never added 3rd party code to their MATLAB path (issue [here](https://github.com/piermorel/gramm/issues/56)

> - [ ]  **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

I could not find any tests in the repository so I was wondering if you'd considered adding any
(issue [here](https://github.com/piermorel/gramm/issues/57)
