### 2018-02-08 piermorel/gramm#56 @piermorel
I will add a few simple instructions for install (maybe where to click on GitHub and a link to Matlab's add to path documentation?).
