### 2018-02-08 piermorel/gramm#57 @piermorel
Thank you for the comment. Currently, ```examples.m``` acts as a sort of test suite: even if it's presented as an example, it goes through all methods of gramm, and often tests multiple options for each. Some of the examples are also specific test cases for edge possibilities (see dodging examples for ```stat_summary``` and ```stat_boxplot``` in examples.m). By running examples.m I can examine the graphical output and make sure that everything is working as it should (and most of that is visual so it's impossible to test in an automated way).  Should there be more than that ?
