### 2018-02-09 piermorel/gramm#57 @cMadan
It might be good to have a folder with an html file that has all of the outputted figures from the examples, so someone who's never used the toolbox can compare the output of `examples.m` with the expected output? (I assume this isn't identical to the panels shown in the README)
