### 2018-02-09 piermorel/gramm#57 @piermorel
@cMadan There actually is an html/ folder in the repository, which contains all the figures and an index.html page produced by using Matlab's publish function on examples.m. I update it every time it's necessary. The images of the README do come from this html folder, but the folder contains more images than the README as it's the full output of examples.m . I indicated this in the paper in the paragraph where I describe available documentation and help.
