### 2018-02-09 piermorel/gramm#57 @trallard
I have checked the examples in the html folder and they are great, but I am afraid that the README documentation section gest a bit lost.
I suggest changing the `html/examples.html` for [html/examples.html](./html/examples.html) or better yet have the html output to actually generate a webpage on your gh-branch and add the link to it in your README, this way people can actually see the rendered html version in their web browsers
