### 2018-02-13 piermorel/gramm#59 @trallard
Statement of need- add to readme 
I think the summary of the [software paper](https://github.com/openjournals/joss-papers/blob/joss.00568/joss.00568/10.21105.joss.00568.pdf) makes an excellent point on why gramm is needed and how it offers a better visualization environment for MATLAB users. This could probably be incorporated to the gramm repository.

This could be in a paragraph in the Readme... something like **Why gramm?**
