### 2018-02-13 #568 @trallard
### Documentation

> - [ ] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I think the summary of the [software paper](https://github.com/openjournals/joss-papers/blob/joss.00568/joss.00568/10.21105.joss.00568.pdf) makes an excellent point on why gramm is needed and how it offers a better visualization environment for MATLAB users. This could probably be incorporated to the gramm repository (issue [here](https://github.com/piermorel/gramm/issues/59))
