### 2018-02-13 f74c06d
Changes for JOSS revision
- Added contribution guidelines
- Expanded installation instructions
- Added statement of need to readme
- Emphasized role of examples.m file and output
