### 2018-02-13 piermorel/gramm#56 @piermorel
Added in latest commit https://github.com/piermorel/gramm/commit/f74c06deb4944d41509667a102f7cf9b162351eb
