### 2018-02-13 piermorel/gramm#57 @piermorel
I used htmlpreview.github.io in order to make the html/examples.html link visible in the browser. I also made the role of examples.m as test file more apparent in CONTRIBUTING.md in https://github.com/piermorel/gramm/commit/f74c06deb4944d41509667a102f7cf9b162351eb
