### 2018-02-13 piermorel/gramm#59 @piermorel
Good suggestion, I added it in the latest commit https://github.com/piermorel/gramm/commit/f74c06deb4944d41509667a102f7cf9b162351eb
