### 2018-02-13 efaa6f7
Added DOI to "grammar of graphics" citation
