+++ paper/paper.bib
@@ -13,6 +13,8 @@
    isbn = {0-387-98774-6},
    publisher = {Springer-Verlag New York, Inc.},
    address = {New York, NY, USA},
+   doi          = {10.1007/978-1-4757-3100-2},
+   url          = {http://dx.doi.org/10.1007/978-1-4757-3100-2}
   } 
   @misc{michael_waskom_2017_883859,
     author       = {Michael Waskom and
