### 2018-02-13 #568 @piermorel
Thank you for these suggestions @trallard. I just adressed all these comments  in commit https://github.com/piermorel/gramm/commit/f74c06deb4944d41509667a102f7cf9b162351eb and closed the corresponding issues. Do not hesitate to reopen them if not satisfactory. The latest commit https://github.com/piermorel/gramm/commit/efaa6f7b4a8b397532b909a581b3641ce05cd196 also adds the doi for "grammar of graphics".
