### 2018-02-14 e999573
Reorganised sections to make the Readme clearer
