### 2018-02-14 b69ca34
Minor rewording
