### 2018-02-15 #568 @trallard
@cMadan I think we can consider the revision of the package completed for the JOSS submission. @piermorel  has sorted out the issues identified and I am sure this package will continue to evolve since it seems to be already well accepted by the MATLAB community.

What are the next steps to be followed?
