### 2018-03-04 #568 @cMadan
@brian-lau, thank you for your interest, but I think I'll go ahead and accept the submission as-is now.

@trallard, thank you again for your review!

@piermorel, can you provide the DOI for an archived version of the project's code? (I.e., archive the current codebase on Zenodo or FigShare). After that we'll be all set to accept!
