### 2018-03-04 #568 @piermorel
Thank you @cMadan, and thank you @trallard for the useful comments.

Gramm is already on Zenodo since several releases. I think Zenodo provides this DOI that allows to track new releases automatically: https://doi.org/10.5281/zenodo.594625 . If there is the need to provide the specific doi from the reviewed version, the latest release on zenodo corresponds to that.
